+++ src/connectedvmware/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+1.0.1
+++++++
+* Fix bug in `az connectedvmware create` command when RG of machine is different from RG of vCenter.
+* Update help command for vm create to remove non-required parameters.
+* Bug fix in sample resource id format for vcenter
+* Add --private-link-scope flag to `az connectedvmware vm guest-agent enable` command.
+
 1.0.0
 ++++++
 * Add command `az connectedvmware create-from-machines`.